refactor edit to remove cobra dependency and also make it reusable
